// OgarniaczAllegro-Frontend/src/app/allegro-stuff/models/single-order-model.ts
export interface ISeller {
  id: string;
  login: string;
  address: IAddress;
  contactDetails: IContactDetails;
}

export interface IUnitPrice {
  amount: string;
  currency: string;
}

export interface IOriginalUnitPrice {
  amount: string;
  currency: string;
}

export interface IOfferPrice {
  amount: string;
  currency: string;
}

export interface IOriginalOfferPrice {
  amount: string;
  currency: string;
}

export interface IOffer {
  id: string;
  title: string;
  friendlyUrl: string;
  unitPrice: IUnitPrice;
  originalUnitPrice: IOriginalUnitPrice;
  quantity: number;
  type: string;
  imageUrl: string;
  productId: string;
  orderDate: Date;
  gold: boolean;
  pharmacy: boolean;
  auction: boolean;
  toRate: boolean;
  orderOfferId: string;
  showOriginalPrice: boolean;
  offerPrice: IOfferPrice;
  originalOfferPrice: IOriginalOfferPrice;
}
// ...
export interface IDelivery {
  cost: ICost;
  name: string;
  methodId: string;
  id: string;
  address: IAddress2;
  generalDelivery: IGeneralDelivery;
  discount: IDiscount;
  status: string;
  timestamp: Date;
  paymentType: IPaymentType;
}

export interface ITotalCost {
  amount: string;
  currency: string;
}

export interface ITotalCost2 {
  amount: string;
  currency: string;
}

export interface ISummary {
  totalCost: ITotalCost2;
  exchangeRateDate: Date;
  showExchangeRateInformation: boolean;
}

export interface IAmount {
  amount: string;
  currency: string;
}

export interface IPayment {
  id: string;
  provider: string;
  amount: IAmount;
  method: string;
  methodId: string;
  status: string;
  date: Date;
  isInstallments: boolean;
  isMultipleSellers: boolean;
  cardMask: string;
}

export interface IPrimary {
  status: string;
  hint: string;
}

export interface IDetail {
  type: string;
  value: string;
  action: string;
}

export interface IAction {
  type: string;
  enabled: boolean;
  details: IDetail[];
}

export interface IStatus {
  primary: IPrimary;
  traits: string[];
  actions: IAction[];
}

export interface ICoins {
  total: number;
}
// ...
export interface ITimeline {
  status: string;
  waybillId: string;
  details: IDetail2[];
  steps: IStep[];
  pickup: IPickup;
  carrier: ICarrier;
  detailedHistoryAvailable: boolean;
  alert: IAlert;
}
// ...
export interface IRescissions {
  rescissions: IRescission[];
  retails: boolean;
}

export interface ISingleorder {
  id: string;
  purchaseId: string;
  seller: ISeller;
  offers: IOffer[];
  delivery: IDelivery;
  totalCost: ITotalCost;
  summary: ISummary;
  payment: IPayment;
  createDate: Date;
  orderDate: Date;
  serviceCountry: string;
  hiddenInMyOrders: boolean;
  boughtOnDevice: string;
  cancelled: boolean;
  status: IStatus;
  coins: ICoins;
  timelines: ITimeline[];
  isFulfilment: boolean;
  groupId: string;
  rescissions: IRescissions;
}

export interface IPrimary2 {
  status: string;
  hint: string;
}

export interface IDetail3 {
  type: string;
  value: string;
  action: string;
}

export interface IAction2 {
  type: string;
  enabled: boolean;
  details: IDetail3[];
}

export interface IStatus2 {
  primary: IPrimary2;
  traits: string[];
  actions: IAction2[];
}

export interface ITotalCost3 {
  amount: string;
  currency: string;
}

export interface ITotalCostBeforeExchange {
  amount: string;
  currency: string;
}

export interface ISummary2 {
  totalCost: ITotalCost3;
  totalCostBeforeExchange: ITotalCostBeforeExchange;
  exchangeRateDate: Date;
  showExchangeRateInformation: boolean;
}

export interface ISingleOrderGroup {
  groupId: string;
  myorders: ISingleorder[];
  status: IStatus2;
  summary: ISummary2;
}

export interface ISingleOrderGroupMeta {
}

export interface IAllegroSingleOrder {
  myorderGroup: ISingleOrderGroup;
  // myorderGroup.$meta: SingleorderGroupMeta;
}